Convert ExitCodes from static class to enum
Converts the internal ExitCodes static class with const int fields to an
internal enum. This improves type safety, discoverability, and enables
Enum.IsDefined checks for validating known MTP exit codes.

This is safe because:
- ExitCodes is internal and not in any PublicAPI.txt
- const int values are inlined at compile time, so extensions compiled
  against the old class have zero runtime dependency on the ExitCodes type
- All out-of-repo IVT consumers (MSTest.Engine, OpenTelemetry, etc.) do
  not reference ExitCodes at all
- Source-embedded consumers always recompile from the same source

Fixes #7807

+++ src/Platform/Microsoft.Testing.Extensions.Retry/RetryOrchestrator.cs
@@ -298,16 +298,16 @@ internal sealed class RetryOrchestrator : ITestHostExecutionOrchestrator, IOutpu
                 catch (OperationCanceledException) when (processExitedCancellationToken.IsCancellationRequested)
                 {
                     await outputDevice.DisplayAsync(this, new ErrorMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.TestHostProcessExitedBeforeRetryCouldConnect, testHostProcess.ExitCode)), cancellationToken).ConfigureAwait(false);
-                    return ExitCodes.GenericFailure;
+                    return (int)ExitCodes.GenericFailure;
                 }
             }
 
             await testHostProcess.WaitForExitAsync().ConfigureAwait(false);
 
             exitCodes.Add(testHostProcess.ExitCode);
-            if (testHostProcess.ExitCode != ExitCodes.Success)
+            if (testHostProcess.ExitCode != (int)ExitCodes.Success)
             {
-                if (testHostProcess.ExitCode != ExitCodes.AtLeastOneTestFailed)
+                if (testHostProcess.ExitCode != (int)ExitCodes.AtLeastOneTestFailed)
                 {
                     await outputDevice.DisplayAsync(this, new WarningMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.TestSuiteFailedWithWrongExitCode, testHostProcess.ExitCode)), cancellationToken).ConfigureAwait(false);
                     retryInterrupted = true;
@@ -370,7 +370,7 @@ internal sealed class RetryOrchestrator : ITestHostExecutionOrchestrator, IOutpu
 
         if (!thresholdPolicyKickedIn && !retryInterrupted)
         {
-            if (exitCodes[^1] != ExitCodes.Success)
+            if (exitCodes[^1] != (int)ExitCodes.Success)
             {
                 await outputDevice.DisplayAsync(this, new ErrorMessageOutputDeviceData(string.Format(CultureInfo.InvariantCulture, ExtensionResources.TestSuiteFailedInAllAttempts, userMaxRetryCount + 1)), cancellationToken).ConfigureAwait(false);
             }
