+++ src/Platform/Microsoft.Testing.Platform/Hosts/ConsoleTestHost.cs
@@ -108,7 +108,7 @@ internal sealed class ConsoleTestHost(
         {
             requestExecuteStop ??= _clock.UtcNow;
 
-            exitCode = ExitCodes.TestSessionAborted;
+            exitCode = (int)ExitCodes.TestSessionAborted;
             await _logger.LogInformationAsync("Test session canceled.").ConfigureAwait(false);
         }
         finally
@@ -128,7 +128,7 @@ internal sealed class ConsoleTestHost(
                 { TelemetryProperties.RequestProperties.AdapterLoadStop, adapterLoadStop },
                 { TelemetryProperties.RequestProperties.RequestExecuteStart, requestExecuteStart },
                 { TelemetryProperties.RequestProperties.RequestExecuteStop, requestExecuteStop },
-                { TelemetryProperties.HostProperties.ExitCodePropertyName, cancellationToken.IsCancellationRequested ? ExitCodes.TestSessionAborted : exitCode.ToString(CultureInfo.InvariantCulture) },
+                { TelemetryProperties.HostProperties.ExitCodePropertyName, cancellationToken.IsCancellationRequested ? (int)ExitCodes.TestSessionAborted : exitCode.ToString(CultureInfo.InvariantCulture) },
             };
 
             if (statistics is not null)
